+++ NEWS.md
@@ -1,3 +1,9 @@
+# wikilake 0.6
+
+* Bug fixes related to non-standard units
+* Bug fixes related to coordinate parsing
+* Added `lake_clean` function and `clean` argument to `lake_wiki` to standardize output units
+
 # wikilake 0.5
 
 * Add ability to pass a vector of names to lake_wiki (#16)
@@ -16,6 +22,3 @@
 
 * Bug fixes related to unit detection and parsing
 * Added a `NEWS.md` file to track changes to the package.
-
-
-